+++ ThirdPartyAdapters/chartboost/CHANGELOG.md
@@ -1,19 +1,30 @@
 ## Chartboost Android Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use new `AdError` API.
+
+#### Version 8.3.0.0
+- Verified compatibility with Chartboost SDK 8.3.0.
+- Updated the minimum required Google Mobile Ads SDK version to 20.5.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 20.5.0.
+- Chartboost SDK version 8.3.0.
+
 #### Version 8.2.1.0
 - Verified compatibility with Chartboost SDK 8.2.1.
 - Updated the minimum required Google Mobile Ads SDK version to 20.1.0.
 
 Built and tested with:
 - Google Mobile Ads SDK version 20.1.0.
-- Chartboost SDK version 8.2.1
+- Chartboost SDK version 8.2.1.
 
 #### Version 8.2.0.1
 - Updated the minimum required Google Mobile Ads SDK version to 20.0.0.
 
 Built and tested with:
 - Google Mobile Ads SDK version 20.0.0.
-- Chartboost SDK version 8.2.0
+- Chartboost SDK version 8.2.0.
 
 #### Version 8.2.0.0
 - Verified compatibility with Chartboost SDK 8.2.0.
@@ -21,14 +32,14 @@ Built and tested with:
 
 Built and tested with:
 - Google Mobile Ads SDK version 19.5.0.
-- Chartboost SDK version 8.2.0
+- Chartboost SDK version 8.2.0.
 
 #### Version 8.1.0.0
 - Verified compatibility with Chartboost SDK 8.1.0.
 
 Built and tested with:
 - Google Mobile Ads SDK version 19.2.0.
-- Chartboost SDK version 8.1.0
+- Chartboost SDK version 8.1.0.
 
 #### Version 8.0.3.2
 - Added descriptive error codes and reasons for adapter load/show failures.